Split qasm-specific tests into their own test file.
